moved prior passing functionality to Samples
